LayoutTests/platform/blackberry/accessibility/img-aria-button-alt-tag-expected.txt


This test makes sure that a img acting as an aria button still returns its alt tag as the description

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


FAIL imgUIElement.description should be AXDescription: alternate. Was .
PASS successfullyParsed is true

TEST COMPLETE
